football_bet_predictor/requirements.txt
fastapi
uvicorn
pandas
numpy
sqlalchemy
psycopg2-binary
scikit-learn
xgboost
python-dotenv
joblib
requests
